+++ examples/react-web-basic/src/examples/reactquery/ReactQueryExample.tsx
@@ -1,20 +1,11 @@
-import { useQueryClient } from "@tanstack/react-query";
+// example app ReactQueryExample.tsx
+
 import { observer } from "mobx-react-lite";
-import { useState } from "react";
+import { useProductQueryStore } from "./BrqContext";
 import ProductListWithQuery from "./components/ProductListWithQuery";
-import { createProductQueryStore, ProductQueryStore } from "./store/productQueryStore";
 
 const ReactQueryExample = observer(() => {
-    const queryClient = useQueryClient();
-
-    const [store] = useState(() => createProductQueryStore(queryClient) as ProductQueryStore);
-
-    // Optionally ensure query is executed when page loads (the productQueryStore has its
-    // middleware configured to automatically fetch products on store creation so this is
-    // just shown as an example)
-    // useEffect(() => {
-    //     store.queries.products();
-    // }, [store]);
+    const store = useProductQueryStore();
 
     return (
         <>