+++ examples/react-web-basic/src/main.tsx
@@ -1,3 +1,5 @@
+// example app main.tsx
+
 import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
 import { StrictMode } from "react";
 import { createRoot } from "react-dom/client";
@@ -5,6 +7,7 @@ import { BrowserRouter, Route, Routes } from "react-router-dom";
 import App from "./App";
 import HomePage from "./HomePage";
 import BasicExample from "./examples/basic/BasicExample";
+import { ProductQueryStoreProvider } from "./examples/reactquery/BrqContext";
 import ReactQueryExample from "./examples/reactquery/ReactQueryExample";
 import "./index.css";
 
@@ -20,15 +23,17 @@ const queryClient = new QueryClient({
 createRoot(document.getElementById("root")!).render(
     <StrictMode>
         <QueryClientProvider client={queryClient}>
-            <BrowserRouter>
-                <Routes>
-                    <Route path="/" element={<App />}>
-                        <Route index element={<HomePage />} />
-                        <Route path="basic" element={<BasicExample />} />
-                        <Route path="reactquery" element={<ReactQueryExample />} />
-                    </Route>
-                </Routes>
-            </BrowserRouter>
+            <ProductQueryStoreProvider>
+                <BrowserRouter>
+                    <Routes>
+                        <Route path="/" element={<App />}>
+                            <Route index element={<HomePage />} />
+                            <Route path="basic" element={<BasicExample />} />
+                            <Route path="reactquery" element={<ReactQueryExample />} />
+                        </Route>
+                    </Routes>
+                </BrowserRouter>
+            </ProductQueryStoreProvider>
         </QueryClientProvider>
     </StrictMode>
 );