+++ packages/beacon/src/index.ts
@@ -1,3 +1,5 @@
+// @apogeelabs/beacon index.ts
+
 export * from "./types";
 export * from "./store";
 export * from "./compose";