+++ packages/beacon-browserstorage/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @apogeelabs/beacon-browserstorage
 
+## 0.3.1
+
+### Patch Changes
+
+- Updated dependencies
+    - @apogeelabs/beacon@0.4.0
+
 ## 0.3.0
 
 ### Minor Changes