+++ packages/beacon-reactquery/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @apogeelabs/beacon-reactquery
 
+## 0.3.0
+
+### Minor Changes
+
+- Cleanup PoC work
+
+### Patch Changes
+
+- Updated dependencies
+    - @apogeelabs/beacon@0.4.0
+
 ## 0.2.0
 
 ### Minor Changes